+++ pwnlib/tubes/process.py
@@ -133,18 +133,21 @@ class process(tube):
         True
         >>> p.connected('send')
         False
-        >>> p.recvline()
-        b'Hello world\n'
+        >>> p.recvline(drop=True)
+        b'Hello world'
         >>> p.recvuntil(b',')
         b'Wow,'
         >>> p.recvregex(b'.*data')
         b' such data'
-        >>> p.recv()
-        b'\n'
+        >>> p.recv()[-1:] == b'\n'
+        True
         >>> p.recv() # doctest: +ELLIPSIS
         Traceback (most recent call last):
         ...
         EOFError
+
+    .. doctest::
+        :options: +POSIX
 
         >>> p = process('cat')
         >>> d = open('/dev/urandom', 'rb').read(4096)
@@ -289,6 +292,10 @@ class process(tube):
             self.suid = self.uid = None
             self.sgid = self.gid = None
             internal_preexec_fn = None
+            # Expect Windows to use CRLF newlines, but if the user explicitly
+            # set it to something else, don't mess with it.
+            if 'newline' not in context._tls and context.newline == b'\n':
+                self.newline = b'\r\n'
         else:
             # Avoid the need to have to deal with the STDOUT magic value.
             if stderr is STDOUT:
@@ -524,6 +531,9 @@ class process(tube):
 
         Example:
 
+        .. doctest::
+            :options: +POSIX +TODO
+
             >>> p = process('/bin/true')
             >>> p.executable == '/bin/true'
             True
@@ -538,6 +548,9 @@ class process(tube):
         """Directory that the process is working in.
 
         Example:
+
+        .. doctest::
+            :options: +POSIX +TODO
 
             >>> p = process('sh')
             >>> p.sendline(b'cd /tmp; echo AAA')
@@ -687,8 +700,14 @@ class process(tube):
 
         The process can choose to ignore this signal, so proper cleanup
         is only done in :meth:`kill`/:meth:`close`.
+
+        Note: On Windows, there is no SIGTERM signal, so the Win32 API function ``TerminateProcess()`` is called
+        instead, leading to :meth:`terminate()` and :meth:`kill()` being effectively the same.
         
         Examples:
+
+        .. doctest::
+            :options: +POSIX
         
             >>> p = process(['python', '-u', '-c', 'import signal;signal.signal(signal.SIGTERM, lambda signum,frame: (print("sigterm"),exit(0)));print("ready");import time;time.sleep(10)'])
             >>> p.recvline_contains(b'ready')
@@ -945,7 +964,10 @@ class process(tube):
             read, write, execute, private, shared, string
 
         Example:
-      
+
+        .. doctest::
+            :options: +POSIX +TODO
+
             >>> p = process(['cat'])
             >>> p.sendline(b"meow")
             >>> p.recvline()
@@ -996,7 +1018,10 @@ class process(tube):
         # addr = address (alias) = start (alias)
 
         from pwnlib.util.proc import memory_maps
-        raw_maps = memory_maps(self.pid)
+        raw_maps = self.poll() is None and memory_maps(self.pid)
+
+        if not raw_maps:
+            self.error("Could not read maps, process %d has finished", self.pid)
 
         maps = []
         # raw_mapping
@@ -1025,7 +1050,10 @@ class process(tube):
         path_value.
 
         Example:
-            
+
+        .. doctest::
+            :options: +POSIX
+
             >>> p = process(['cat'])
             >>> mapping = p.get_mapping('[stack]')
             >>> mapping.path == '[stack]'
@@ -1068,6 +1096,9 @@ class process(tube):
 
         Example:
 
+        .. doctest::
+            :options: +POSIX
+
             >>> p = process(['cat'])
             >>> mapping = p.stack_mapping()
             >>> mapping.path
@@ -1096,6 +1127,9 @@ class process(tube):
         Returns :meth:`.process.get_mapping` with '[heap]' and single as arguments.
 
         Example:
+
+        .. doctest::
+            :options: +POSIX
 
             >>> p = process(['cat'])
             >>> p.sendline(b'meow')
@@ -1129,6 +1163,9 @@ class process(tube):
 
         Example:
 
+        .. doctest::
+            :options: +LINUX
+
             >>> p = process(['cat'])
             >>> mapping = p.vdso_mapping()
             >>> mapping.path
@@ -1157,6 +1194,9 @@ class process(tube):
         Returns :meth:`.process.get_mapping` with '[vvar]' and single as arguments.
 
         Example:
+
+        .. doctest::
+            :options: +LINUX
 
             >>> p = process(['cat'])
             >>> mapping = p.vvar_mapping()
@@ -1187,6 +1227,9 @@ class process(tube):
         or all libc mappings, depending on "single". 
 
         Example:
+
+        .. doctest::
+            :options: +POSIX
 
             >>> p = process(['cat'])
             >>> p.sendline(b'meow')
@@ -1268,6 +1311,9 @@ class process(tube):
 
         Example:
 
+        .. doctest::
+            :options: +POSIX
+
             >>> p = process(['cat'])
             >>> p.sendline(b'meow')
             >>> p.recvline()
@@ -1306,7 +1352,9 @@ class process(tube):
 
         Example:
 
-            >>> from pwn import *
+        .. doctest::
+            :options: +POSIX
+
             >>> p = process(['cat'])
             >>> p.send(b'meow')
             >>> p.recvuntil(b'meow')
@@ -1346,6 +1394,9 @@ class process(tube):
 
         Example:
 
+        .. doctest::
+            :options: +POSIX
+
             >>> p = process(['cat'])
             >>> p.sendline(b'meow')
             >>> p.recvline()
@@ -1377,17 +1428,11 @@ class process(tube):
         by the process to the address it is loaded at in the process' address
         space.
         """
-        maps_raw = self.poll() is None and self.maps()
-
-        if not maps_raw:
-            import pwnlib.elf.elf
-
-            with context.quiet:
-                return pwnlib.elf.elf.ELF(self.executable).maps
+        all_maps = self.maps()
 
         # Enumerate all of the libraries actually loaded right now.
         libs = {}
-        for mapping in maps_raw:
+        for mapping in all_maps:
             path = mapping.path
             if os.sep not in path: continue
             path = os.path.realpath(path)
@@ -1406,16 +1451,19 @@ class process(tube):
 
         Example:
 
-        >>> p = process("/bin/cat")
-        >>> p.send(b"meow")
-        >>> p.recvuntil(b"meow")
-        b'meow'
-        >>> libc = p.libc
-        >>> libc is not None
-        True
-        >>> libc # doctest: +SKIP
-        ELF('/lib64/libc-...so')
-        >>> p.close()
+        .. doctest::
+            :options: +POSIX
+
+            >>> p = process("/bin/cat")
+            >>> p.send(b"meow")
+            >>> p.recvuntil(b"meow")
+            b'meow'
+            >>> libc = p.libc
+            >>> libc is not None
+            True
+            >>> libc # doctest: +SKIP
+            ELF('/lib64/libc-...so')
+            >>> p.close()
         """
         from pwnlib.elf import ELF
 
@@ -1493,6 +1541,9 @@ class process(tube):
 
         Example:
 
+        .. doctest::
+            :options: +POSIX +TODO
+
             >>> e = ELF(which('bash-static'))
             >>> p = process(e.path)
 
@@ -1528,6 +1579,9 @@ class process(tube):
         Example:
         
             Let's write data to  the beginning of the mapped memory of the  ELF.
+
+        .. doctest::
+            :options: +POSIX +TODO
 
             >>> context.clear(arch='i386')
             >>> address = 0x100000