+++ pwnlib/windbg.py
@@ -3,10 +3,23 @@ During exploit development, it is frequently useful to debug the
 target binary under WinDbg. This module provides a simple interface
 to do so under Windows.
 
+By default, :attr:`.context.debugger` is set to ``"auto"``, which will
+attempt to automatically select an appropriate debugger based on the
+available debuggers on the system.
+
+The order of preference is:
+- ``windbgx``
+- ``windbg``
+
+If automatic lookup fails, you can manually set :attr:`.context.debugger` to
+the debugger of your choice and provide the path to the debugger binary
+using :attr:`.context.windbgx_binary` or :attr:`.context.windbg_binary`.
+
 Useful Functions
 ----------------
 
 - :func:`attach` - Attach to an existing process
+- :func:`debug` - Start a new process under the debugger
 
 Debugging Tips
 --------------
@@ -61,6 +74,7 @@ import os
 import signal
 
 import subprocess
+import tempfile
 
 from pwnlib import tubes
 from pwnlib.context import LocalContext
@@ -74,14 +88,14 @@ log = getLogger(__name__)
 CREATE_SUSPENDED = 0x00000004
 
 @LocalContext
-def debug(args, windbgscript=None, exe=None, env=None, creationflags=0, **kwargs):
-    """debug(args, windbgscript=None, exe=None, env=None, creationflags=0) -> tube
+def debug(args, dbgscript=None, exe=None, env=None, creationflags=0, **kwargs):
+    """debug(args, dbgscript=None, exe=None, env=None, creationflags=0) -> tube
 
     Launch a process in suspended state, attach debugger and resume process.
 
     Arguments:
         args(list): Arguments to the process, similar to :class:`.process`.
-        windbgscript(str): windbg script to run.
+        dbgscript(str): windbg script to run.
         exe(str): Path to the executable on disk.
         env(dict): Environment to start the binary in.
         creationflags(int): Flags to pass to :func:`.process.process`.
@@ -96,10 +110,10 @@ def debug(args, windbgscript=None, exe=None, env=None, creationflags=0, **kwargs
             # Create a new process, and stop it at 'main'
             io = windbg.debug('calc', '''
             bp $exentry
-            go
+            g
             ''')
 
-        When WinDbg opens via :func:`.debug`, it will initially be stopped on the very first
+        When the debugger opens via :func:`.debug`, it will initially be stopped on the very first
         instruction of the entry point.
     """
     if isinstance(
@@ -111,43 +125,88 @@ def debug(args, windbgscript=None, exe=None, env=None, creationflags=0, **kwargs
         log.warn_once("Skipping debugger since context.noptrace==True")
         return tubes.process.process(args, executable=exe, env=env, creationflags=creationflags)
     
-    windbgscript = windbgscript or ''
-    if isinstance(windbgscript, str):
-        windbgscript = windbgscript.split('\n')
+    dbgscript = dbgscript or ''
+    if isinstance(dbgscript, str):
+        dbgscript = dbgscript.split('\n')
     # resume main thread
-    windbgscript = ['~0m'] + windbgscript
+    dbgscript = ['~0m'] + dbgscript
     creationflags |= CREATE_SUSPENDED
     io = tubes.process.process(args, executable=exe, env=env, creationflags=creationflags)
-    attach(target=io, windbgscript=windbgscript, **kwargs)
+    attach(target=io, dbgscript=dbgscript, **kwargs)
 
     return io
 
 def binary():
     """binary() -> str
 
-    Returns the path to the WinDbg binary.
+    Returns the path to the debugger binary depending on the context.
+    :attr:`.context.debugger` is used to determine which debugger to use.
 
     Returns:
         str: Path to the appropriate ``windbg`` binary to use.
     """
-    windbg = misc.which('windbgx.exe') or misc.which('windbg.exe')
-    if not windbg:
-        log.error('windbg is not installed or in system PATH')
-    return windbg
+    if context.debugger == 'auto':
+        for debugger in context.debugger_choices:
+            with context.local(debugger=debugger, log_level='critical'):
+                try:
+                    return binary()
+                except Exception:
+                    pass
+        else:
+            log.error('No debugger found. Please set context.debugger to one of: %s\n'
+                      'You might have to specify the path to the debugger binary with context.windbg_binary or context.windbgx_binary.',
+                      ', '.join(context.debugger_choices))
+    
+    if context.debugger == 'windbg':
+        if context.windbg_binary:
+            windbg = misc.which(context.windbg_binary)
+            if not windbg:
+                log.warn_once('Path to WinDbg binary `{}` not found'.format(context.windbg_binary))
+            return windbg
+
+        windbg = misc.which('windbg.exe')
+        if not windbg and os.environ.get('ProgramFiles(x86)'):
+            arch_str = {
+                'i386': 'x86',
+                'amd64': 'x64',
+                'aarch64': 'arm64',
+            }.get(context.arch)
+            if not arch_str:
+                log.error('Unsupported architecture for windbg: {}'.format(context.arch))
+            windbg = os.path.join(os.environ.get('ProgramFiles(x86)'), 'Windows Kits', '10', 'Debuggers', arch_str, 'windbg.exe')
+        if not windbg or not os.path.exists(windbg):
+            log.error('windbg is not installed or in system PATH. You can set context.windbg_binary to specify the path manually.')
+        return windbg
+
+    if context.debugger == 'windbgx':
+        if context.windbgx_binary:
+            windbg = misc.which(context.windbgx_binary)
+            if not windbg:
+                log.warn_once('Path to WinDbgx binary `{}` not found'.format(context.windbgx_binary))
+            return windbg
+
+        windbg = misc.which('windbgx.exe')
+        if not windbg and os.environ.get('LocalAppData'):
+            windbg = os.path.join(os.environ.get('LocalAppData'), 'Microsoft', 'WindowsApps', 'WinDbgX.exe')
+        if not windbg or not os.path.exists(windbg):
+            log.error('windbgx is not installed or in system PATH. You can set context.windbgx_binary to specify the path manually.')
+        return windbg
+
+    log.error('Invalid debugger selection: %s', context.debugger)
 
 @LocalContext
-def attach(target, windbgscript=None, windbg_args=[]):
-    """attach(target, windbgscript=None, windbg_args=[]) -> int
+def attach(target, dbgscript=None, dbg_args=[]):
+    """attach(target, dbgscript=None, dbg_args=[]) -> int
 
     Attach to a running process with WinDbg.
 
     Arguments:
         target(int, str, process): Process to attach to.
-        windbgscript(str, list): WinDbg script to run after attaching.
-        windbg_args(list): Additional arguments to pass to WinDbg.
+        dbgscript(str, list): Debugger script to run after attaching.
+        dbg_args(list): Additional arguments to pass to the debugger.
 
     Returns:
-        int: PID of the WinDbg process.
+        int: PID of the debugger process.
 
     Notes:
 
@@ -173,7 +232,7 @@ def attach(target, windbgscript=None, windbg_args=[]):
         Attach a debugger to a :class:`.process` tube and automate interaction
 
         >>> io = process('cmd') # doctest: +SKIP
-        >>> pid = windbg.attach(io, windbgscript='''
+        >>> pid = windbg.attach(io, dbgscript='''
         ... bp kernelbase!WriteFile
         ... g
         ... ''') # doctest: +SKIP
@@ -204,33 +263,37 @@ def attach(target, windbgscript=None, windbg_args=[]):
         log.error('could not find target process')
     
     cmd = [binary()]
-    if windbg_args:
-        cmd.extend(windbg_args)
+    if dbg_args:
+        cmd.extend(dbg_args)
     
     cmd.extend(['-p', str(pid)])
 
-    windbgscript = windbgscript or ''
-    if isinstance(windbgscript, str):
-        windbgscript = windbgscript.split('\n')
-    if isinstance(windbgscript, list):
-        windbgscript = ';'.join(script.strip() for script in windbgscript if script.strip())
-    if windbgscript:
-        cmd.extend(['-c', windbgscript])
+    dbgscript = dbgscript or ''
+    if isinstance(dbgscript, str):
+        dbgscript = dbgscript.split('\n')
+    with tempfile.NamedTemporaryFile(mode='w+', delete=False, suffix='.dbg') as tmp:
+        tmp.write('\n'.join(script.strip() for script in dbgscript if script.strip()))
+        tmp.flush()
+        dbgscript_file = tmp.name
+    
+    if dbgscript:
+        cmd.extend(['-c', '$<{}'.format(dbgscript_file)])
     
     log.info("Launching a new process: %r" % cmd)
 
     io = subprocess.Popen(cmd)
-    windbg_pid = io.pid
+    debugger_pid = io.pid
 
     def kill():
         try:
-            os.kill(windbg_pid, signal.SIGTERM)
+            os.unlink(dbgscript_file)
+            os.kill(debugger_pid, signal.SIGTERM)
         except OSError:
             pass
 
     atexit.register(kill)
 
     if context.native:
-        proc.wait_for_debugger(pid, windbg_pid)
+        proc.wait_for_debugger(pid, debugger_pid)
 
-    return windbg_pid
+    return debugger_pid